Extract a shared profile picture picker modal
Lift the accounts profile-picture modal body (preview header, option grid
with hover/focus ring and check badge, modal actions with submit/error
handling) into core AppProfilePicturePickerModal, parameterized by title,
keys, and selection semantics. Accounts keeps a thin wrapper with its
existing keys and behavior.
